https://sqlinfocg.cegepgranby.qc.ca/2147186
Mettre tout les commits de dev dans main

Related work items: #97

+++ VLISSIDES/VLISSIDES/ViewModels/Accueil/LivreCardVM.cs
@@ -42,7 +42,7 @@ public class LivreCardVM
 
     [DisplayName("Auteurs")] public string Auteurs { get; set; }
 
-    [DisplayName("Score")] public int Score { get; set; }
+    [DisplayName("Score")] public double Score { get; set; }
 
     [DisplayName("Catégories associés")] public string Categorie { get; set; }
 }
